Add files via upload
This is my project which i have created in my collage. which helps us to identification(biometric), where we track the face and capture the image, captured image is stored in the binary file format. And this system can be  mostly used in the attendance,   in police or government services  which can track the employees leaves and present time.