Add files via upload
I have created this project by using Django, python, CSS, and html. where the google services can be accessed . in gmail.html this file contain the code of the gmail and map.html contain map code. where as show. html contain the "Google services" message and both the file as a link attached on it.